refactor hook->hooks
