+++ src/lib.rs
@@ -2,7 +2,7 @@ pub mod capabilities;
 pub mod commands;
 pub mod container;
 pub mod dbus;
-pub mod hook;
+pub mod hooks;
 pub mod logger;
 pub mod namespaces;
 pub mod notify_socket;
